+++ bootstrap/providers.php
@@ -3,4 +3,5 @@
 return [
     App\Providers\AppServiceProvider::class,
     App\Shared\Providers\ModuleServiceProvider::class,
+    App\Shared\Providers\AutorizacionServiceProvider::class,
 ];